+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,11 @@ Release History
 upcoming
 ++++++
 * 'az containerapp debug': Open an SSH-like interactive shell within a container app debug console.
+* 'az containerapp create': New target-label option for Labels revision mode.
+* 'az containerapp update': New revisions-mode and target-label options for Labels revision mode.
+* 'az containerapp up': New revisions-mode and target-label options for Labels revision mode.
+* 'az containerapp label-history': New list and show commands to see label revision assignment history.
+* 'az containerapp revision set-mode': New target-label option for Labels revision mode.
 
 1.1.0b1
 ++++++
@@ -13,7 +18,7 @@ upcoming
 * 'az containerapp create/update': `--yaml` support property pollingInterval and cooldownPeriod
 * 'az containerapp session code-interpreter upload-file/list-files/show-file-content/show-file-metadata/delete-file': Support `--path` to specify the path of code interpreter session file resource
 * 'az containerapp session code-interpreter': Update response payload format for api-version 2024-10-02-preview
-* 'az containerapp env maintenance-config add/update/list/remove': Support environment maintenance config management 
+* 'az containerapp env maintenance-config add/update/list/remove': Support environment maintenance config management
 * 'az containerapp sessionpool create': Support managed identity when create session pool with --mi-system-assigned --mi-user-assigned
 
 1.0.0b4